+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   * [FIX]: (Selfie) Add http:// protocol to links if user doesn't supply them
   * [FEATURE]: (Selfie) Default the call to action to the Website link if available
   * [FEATURE]: (Selfie) Add Social Sharing link in Campaign Creation screen
+  * [FEATURE]: (Selfie) Add ability to manage campaigns (pause, resume, cancel)
 
 ## v1.23.0 (October 27, 2015)
 * *[v1.23.0.rc3]*